Conflicts:
	src/main/example-data/DataSources.xml
	src/main/java/de/tuberlin/dima/presslufthammer/transport/QueryHandler.java
	src/main/java/de/tuberlin/dima/presslufthammer/xml/DataSourcesReaderImpl.java
	src/main/java/scratch/QueryTest.java

+++ src/main/java/de/tuberlin/dima/presslufthammer/transport/util/QueryHandler.java
@@ -39,13 +39,13 @@ public class QueryHandler {
 	final Channel client;
 	final QueryMessage queryMsg;
 	final Query query;
-	private int numPartsExpected;
+	private long numPartsExpected;
 	QueryStatus status;
 	SchemaNode schema;
 	List<InMemoryReadonlyTablet> parts;
 	long finished = 0L;
 
-	public QueryHandler(int parts, QueryMessage queryMsg, SchemaNode schema,
+	public QueryHandler(long parts, QueryMessage queryMsg, SchemaNode schema,
 			Channel client) {
 		assert (queryMsg.getQueryId() > 0);
 		this.parts = Lists.newLinkedList();
